ADD: Add placeholder comment message when a comment is submitted.
It's weird that there is no feedback on submitting a comment if the site is set to approve comments before showing up. This lets you know it's not just lost in the ether, while still being simpler than checking through the returned thread on submittal to see if the new comment is added.